Moves paginator sequences under `TCClient.Paginator` namespace
